+++ pkg/providers/property/testdata/TestResourceEdgeHostname/new_akamaized_error_update_ipv6_performance.tf
@@ -7,10 +7,7 @@ resource "akamai_edge_hostname" "edgehostname" {
   group               = "grp_2"
   product             = "prd_2"
   edge_hostname       = "test.akamaized.net"
-  ip_behavior         = "IPV6_COMPLIANCE"
-  status_update_email = []
+  ip_behavior         = "IPV6_PERFORMANCE"
+  status_update_email = ["hello@akamai.com"]
 }
 
-output "edge_hostname" {
-  value = akamai_edge_hostname.edgehostname.edge_hostname
-}